OCR model 2 & 3/requirements.txt
flask
torch
transformers
Pillow
fpdf
python-docx
pdf2image
werkzeug
